requirements.txt
deepgram_sdk==3.2.3
Flask==2.3.2
openai==1.14.3
